requirements.txt
# Discord Bot Requirements
# Core dependencies for the Monstrum Discord Bot

# Discord.py - Main Discord API wrapper
discord.py>=2.3.0

# Additional useful packages
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# Development and logging
colorlog>=6.7.0

# For future database integration (optional)
aiosqlite>=0.19.0

# For configuration management (optional)
python-dotenv>=1.0.0

# For web scraping game news (optional)
beautifulsoup4>=4.12.0
requests>=2.31.0